Move `gitutil`.
