+++ analytics/analytics.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/earthly/earthly/cliutil"
 	"github.com/earthly/earthly/fileutil"
-	"github.com/earthly/earthly/gitutil"
+	"github.com/earthly/earthly/util/gitutil"
 	"github.com/earthly/earthly/util/syncutil"
 
 	"github.com/google/uuid"
