+++ buildcontext/local.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 
 	"github.com/earthly/earthly/domain"
-	"github.com/earthly/earthly/gitutil"
+	"github.com/earthly/earthly/util/gitutil"
 	"github.com/earthly/earthly/util/llbutil"
 	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/earthly/earthly/util/syncutil/synccache"
