+++ util/gitutil/detectgit.go
