+++ util/gitutil/detectgit_test.go
